Add .ruby-version guideline
The `.ruby-version` file is a common standard used by Ruby switchers to
choose which version of Ruby to be used by the project. Similar to the
`Gemfile.lock` file, which locks gem dependencies, this file helps
maintain development-production parity and avoid bugs such as:

https://twitter.com/sstephenson/status/286599900932739072

While the patch level theoretically should not make a difference, the
reality is that Ruby has a history of changing things with patch
releases. For example, Enumerator (and related methods) were introduced
in a patch level of 1.8.7.