Prefix branches with initials when solo
When pairing, it is common to rotate pairs frequently. Often, the two
people that begin a branch are not the same people that continue working
on the branch. This leads to the branch name lying to everyone.

Therefore, I suggest that we don't need to add the initials to the
branch name when pairing.

One could argue that branches shouldn't last long enough to support a
pair swap. If that's true, then initial prefixing should be completely
unnecessary.