Add "Accepting a GitHub Pull Request" protocol
The only potentially controversial part of this protocol is the `co-pr` alias.

In general, we've avoided using aliases and non-standard scripts when describing
our protocol. For example, in the [Git Protocol], commands such as `git fetch
origin` and `git rebase -i origin/master` are spelled out, even though just
about everybody uses an alias for these steps.

[Git Protocol]: https://github.com/thoughtbot/guides/tree/master/protocol/git

My argument for showing the alias here is that I don't think it's very pragmatic
to have the pull request number repeated three times in the guide:

    git fetch origin pull/123/head:pr/123
    git checkout pr/123

As programmers, we're all likely to use an alias for this. I think the right
balance is to define the alias in the guide, and then use it.