Avoid backticks in CoffeeScript
Backticks allow snippets of JavaScript to be embedded in CoffeeScript. While
some folks consider backticks useful in a few niche circumstances, they should
be avoided because so none of JavaScript's "bad parts", like <tt>with</tt> and
<tt>eval</tt>, sneak into CoffeeScript.

This can be enforced by Hound with the `no_backticks` option in CoffeeLint,
which is enabled by default.

http://www.coffeelint.org/#options

This is an "Avoid", not a "Don't", because in `ember-cli` the import statements
do not work in CoffeeScript and we have to use the backticks:

http://www.ember-cli.com/#coffeescript

There are some cases where it is unavoidable.

Also: alphabetize the CoffeeScript guidelines.