Add guideline for creating mutable collections
The alternative syntax is:

```swift
var collection = [Object]()
```

I find this syntax jarring because of the visual combination of the
braces and the parens. It also seems to be taking a too-rigid stance in
favor of using type inference. I believe that the proposed syntax lets
you type the object, and then be explicit about creating an empty array,
which will then by typed appropriately.