Add style guide for order of properties in Android views
Alphabetized is pretty much the standard, but we felt that `id`,
`layout_width`, and `layout_height` should be listed first. `id` is
the property you care about most (if it is present), and should be
quick to find at a glance. `layout_width` and `layout_height` are
required on all views, and affect the layout of the view most, so they
should also be separated from the rest.