Only follow a best practice if you understand it
The styles should always be followed, and anything that is a style
guideline should be defensible with just "we picked this style."

The best practices are things that we've learned through experience and
teaching, and therefore cannot be conveyed completely in a pithy
guideline. They are reminders for those who understand them, not lessons
for those looking to learn.

As such, if you do not understand a best practice guideline, then you
are simply not the target audience.