A pull system means never assigning
This introduces a new communications section for describing how best to
communicate like a thoughtbotter.

Up first: tickets. We use Trello right now, but the principles of how we
use it are the same as when we used Pivotal Tracker, JIRA, Trajectory,
GitHub Issues, and everything else. Therefore I've left it as just
"tickets."

We follow a pull system, which means:

* For client projects, the next designer or developer takes the ticket
  with top priority.
* For discussion topics, such as happens on the Research board, people
  subscribe to tickets they are interested in.

We use a modified Kanban/CONWIP at thoughtbot. This commit message is
not the place to go into the full details of why we chose to do that,
but at its simplest the output is controlled by demand (e.g. "the
user wants this" or "the blog post author wants to write this") and
not by forecast ("this is due on Thursday", or by analogy "I want you to
write this"). This reflects how we believe software can be built and
how to run a sustainable, happy workplace.