Decide when to merge your branch
The value of a code review is in the feedback from other developers.
Once the developer gets feedback, it's up to hir to decide whether the
branch is ready to merge.

Sign-off is handy for some repos which affect everyone (guides,
playbook, handbook, etc.). Others, such as the blog or code, can be more
lax. As such, simplify the suggestion in the guide and empower the
merger to take power.