Break up style guide by language/framework
Our Styles guideline has grown large and cumbersome. It is over 400
lines long and has 22 sections (some of which have several subsections.
The directory structure looks like this:

    style
    ├── README.md
    └── samples
        ├── ObjectiveC.m
        ├── Swift.swift
        ├── acceptance_test.rb
        ├── android_layout.xml
        ├── backbone.coffee
        ├── coffee.coffee
        ├── erb.rb
        ├── haskell.hs
        ├── migration.rb
        ├── predicate_tests.rb
        ├── ruby.rb
        ├── sass.scss
        └── testing.rb

In this commit, I've left some of the general section in the main README
and broken out all of the language/framework-specific stuff into
subdirectories, each with their own README and examples. The directory
structure now looks like:

    style
    ├── README.md
    ├── android
    │   ├── README.md
    │   └── android_layout.xml
    ├── backbone
    │   ├── README.md
    │   └── sample.coffee
    ├── coffeescript
    │   ├── README.md
    │   └── sample.coffee
    ├── ember
    │   └── README.md
    ├── erb
    │   ├── README.md
    │   └── sample.html.erb
    ├── git
    │   └── README.md
    ├── haskell
    │   ├── README.md
    │   └── sample.hs
    ├── html
    │   └── README.md
    ├── objective_c
    │   ├── README.md
    │   └── sample.m
    ├── python
    │   └── README.md
    ├── rails
    │   ├── README.md
    │   └── migration.rb
    ├── ruby
    │   ├── README.md
    │   └── sample.rb
    ├── sass
    │   ├── README.md
    │   └── sample.scss
    ├── shell
    │   └── README.md
    ├── swift
    │   ├── README.md
    │   └── sample.swift
    └── testing
        ├── README.md
        ├── acceptance_test_spec.rb
        ├── predicate_tests_spec.rb
        └── unit_test_spec.rb

This should make it easier to look up the relevant style guide as well
as make it simpler to deal with multiple, smaller sample files for each
language. This organization pattern would also support
https://github.com/thoughtbot/guides/issues/261 very well.

This commit simply changes the directory structure and breaks out
sections of the style guide. Other refactorings that could build on this
are:

* Break up long lists of bullet points under several headings
* Break up large example files into smaller more focused example files