Prefer `+` over `\` for concatenating long strings
In Ruby, `"a" "b" == "ab"`, i.e. whitespace can be used to concatenate strings.
Due to this, the following can be used to concatenate strings:

    "hello" \
    "world"

It is equivalent to `"hello" "world"`.

But we don't ever write `"hello" "world"`, we write `"hello" + "world"` and rely
on `String#+` instead of niche syntax. Therefore we should use `+` to
concatenate multi-line strings as well:

    "hello" +
      "world"