Ember: Prefer `ember-fetch` over `ic-ajax`
Stefan is looking for people to try this out and give feedback.
We have enough Ember projects that we could contribute
good information to this RFC and the spec.

There isn't much risk to establishing this preference
because it's being polyfilled so it will work wherever.

The gain is that we will be able to
provide valuable feedback to the ember community.
Assuming this eventually get's merged we will also be ahead
of the curve and not need to update existing apps
to accomidate it.

> What is ember-fetch?
>
> * a wrapper around: https://github.com/github/fetch
> * which is a partial polyfil for  https://fetch.spec.whatwg.org/
> * more info: https://github.com/stefanpenner/ember-fetch
> * never falls back to native (so always works the same, and works with pretender)
>
> Why is this discussion happening now and not earlier?
>
> * ember 2 no longer supports IE8, and github/fetch is IE9+
> * fetch is a spec primitive of the DOM, it stands to reason we should prefer that over $.ajax
> * maybe we can catch issues to help adapt/improve/add to the spec
> * $.ajax is crazy quirky
> * uploads have progress in the request object (unsure if the polifill does this or not actually, need to verify)
>
> Cons:
>
> * its not $.ajax which, although not a spec thing, is basically a primitive many devs are used to
> * features like `$.ajaxPrefilter` are not obviously a thing in `fetch` (maybe some other approach ?)
> * .... ?
>
>
> Noop:
>
> * still no formal cancelation
>
> RFC: https://github.com/ember-cli/rfcs/issues/19