Adjust line-wrapping rules for Rails macros
See slack conversation:
https://thoughtbot.slack.com/archives/ruby/p1444250120000353

This adds a new rule for Rails that overrides the default line break
rule in the main README.

This aims to make it easier to scan lists of macros and find the one
you're looking for without having to look at the following line if it's
wrapped.

See the new examples.