Remove argument list style guide
This guideline has proven to be the most unintuitive one, most often
broken and complained about in Slack and real life. It's the one that we
only get right after Hound complains other times. In short, it's not
natural.

Removing this guideline allows us to put the art back into formatting,
giving us the freedom to align arguments in the style that is most
readable and expressive.