Suggest referencing class by constant in AR macros
For ActiveRecord associations where AR can't automatically determine the
class for the association, the `class_name` option allows the developer
to specify what class should be used. This option can take either the
constant of the model class `Person`, or a string that matches the
class' name `"Person"`.

Using a string to specify this option postpones feedback that the class
might be missing. For example, if my model has an association like this
where the `Author` class does not yet exist:

```ruby
    class Post < ActiveRecord::Base
      has_one :author, class_name: "Person"
    end
```

Specs will only fail if they interact with the association. If
`class_name` is specified as `class_name: Person`, and `Person` does not
exist, an uninitialized constant `NameError` exception will stop all
execution of the application.

There may possibly be some memory-use benefit to creating fewer strings
during application initialization, but I think the early failure
feedback benefit of this practice provides enough value on its own to
recommend it as a best practice.