Specify when to use () and Void in func signatures
We want to avoid ambiguity, especially with our return types. Functions
that return `()` can become hard to read when passed as an argument to a
higher order function. For example:

```swift
func f(g: () -> ()) { }
```

The parens for the return value end up getting lost in the closing
parens for the higher order function. Conversely, by using `Void`:

```swift
func f(g: () -> Void) { }
```

`Void` stands out as a specific type on its own, and doesn't get lost in
a mess of parens.

At the same time, `()` is a nice brief indication that a function takes
no arguments, and mimics the look of an empty tuple. So if you declare
these two functions:

```swift
func fa(g: () -> Void) { }
func fb(g: (a: T) -> Void) { }
```

There is a visual similarity between the two functions.

This is also how Apple prefers to style functions taking or returning
`Void`.