Use 2 spaces for indentation in Swift
This brings Swift inline with the defaults for Ruby and other languages.
The benefit here is that while working outside of Xcode (as some of us
try to do), we can use a consistent setting for indentation instead of
having to special case Swift.

From a visual aspect, I've gotten used to the look of 2 spaces, and I
think it feels nice to move away from the 4 space indentations of
Objective-C.

A few of our projects (internal and external) are already using 2
spaces, and I think in _general_ people have liked it more.

Note that this has already been proposed, but was shot down because we
couldn't come to a real consensus. I'd like to re-poll the team and see
where we stand now.