Advocate using locals to render views
It has [long] been [argued] that using instance variables to pass values
to Rails views isn't the greatest of ideas: Rails twists instance
variables to do something they aren't supposed to do, and if you think
about it, it's unclear how views even have access to variables that live
in a completely different scope. (Hint: they're copied from the
controller to the view at runtime when the view is instantiated.)

In most other languages, views (or templates, as they're sometimes
called) are essentially functions and, as such, when you're rendering a
template and want to give that template access to certain values, you
just pass those values along as arguments.

Is there a way to do this in Rails? Yes, by calling `render` explicitly
and using the `locals` option. So your controller would go from this:

``` ruby
class PostsController < ApplicationController
  def index
    @posts = Post.all
  end
end
```

to this:

``` ruby
class PostsController < ApplicationController
  def index
    posts = Post.all
    render locals: { posts: posts }
  end
end
```

The upside to this approach is that now, views and partials act the same
in terms of how they reference variables, and you can convert a view to
a partial easily.

The downside is that you can no longer easily assert that a view was
rendered with a particular variable, but that's okay, because we don't
write controller specs anyway.

[long]: https://github.com/hashrocket/decent_exposure
[argued]: http://thepugautomatic.com/2013/05/locals/