Reorganize Swift sample code
This is kind of a weird change, but this file is fairly unruly and I'm
not sure how best to structure it so that it's useful. I ended up
removing a couple of things that I didn't think were needed and changed
up the sections to be a bit more structured.