Start a quick "how to ..." section
This section is to address the growing amount of internal tools we have
built over time; it is an attempt at making them more discoverable in a
problem/solution format instead of a long list, plus a quick-start
guide.

This is in RST format so that we can make use of the table of contents
feature.