Place methods and tests for methods in same order
When writing tests for a class, it's nice to have the methods and the
tests for those methods listed in the same order. Why? Not because it
makes it easier to find tests (because you can simply use Ctrl-F) or
easier to know where to put new tests (because you could always put them
at the bottom). Really, it's the same argument behind alphabetizing
associations -- it just looks cleaner, and shows attention to detail.