Swift: Force unwrap `segue.destinationViewController`
We should force unwrap the `destinationViewController` on a segue when it
matches one of our identifier.

This causes us to fail as early and loud as possible when
something unexpected happens.

Alternatives considered
-----------------------

**Optional unwrap or guard let/return**:

While this is safer, it will silently not do anything when this isn't a case we
expected. If we're expecting a specific type of vc and it's not that type, we
should be alerted somehow so we can fix it.

**Guard let/fatalError**:

I don't have much of an argument against this one other than it's more verbose.
I don't think the verbosity would improve debugging much so I like the less
verbose version.