Swift: Prefer presenting a view controller manually
In the discussion about force unwrapping (#430)
`UISegue#destinationViewController` we got sidetracked in talking about
dropping segues all together. This is to continue that discussion
without it intervening with the discussion about force unwrapping.

I feel segues can cause more issues than they're worth. The value in
storyboards is not in segues, but in the visual layout. Segues are
annoying because they pretty much guarantee having to keep state around.
If a value needs to be communicated to the destination of a segue, that
needs to happen in `prepareForSegue`. This decouples presenting the view
controller from configuring it.

By dropping segues we can manually instantiate and configure view
controllers. Initialization can be done using
`UIStoryboard#instantiateViewControllerWithIdentifier`.