Prefer curly braces in Ruby
The big difference between `do`/`end` and `{`/`}`, with regards to
blocks, is [precedence].

As a simplification, let's pick one and always use it. How about curly
braces (`{`/`}`) as that one?

Due to precedence, this will not work appropriately for `#it` as we
currently write it. We could use parens around `#it`'s argument:

```ruby
it("redirects to the home page") {
  get :new
  expect(response).to redirect_to(root_url)
}
```

Or we could make it a "prefer" instead of "use", allowing for the
responsible freedom that comes with Ruby's syntax, but clearly stating
our preference.

[precedence]: https://robots.thoughtbot.com/do-end-your-braces