Fix warnings and errors in rubocop.yml
Rubocop v0.50.0 moved some cops and renamed them. This means that prior
to this commit, running rubocop would result in a number of warnings
and fail on an error.

To fix the warnings:

- Rename `Style/AccessorMethodName` to `Naming/AccessorMethodName`
- Rename `Style/AsciiIdentifiers` to `Naming/AsciiIdentifiers`
- Rename `Style/FileName` to `Naming/FileName`
- Rename `Style/PredicateName` to `Naming/PredicateName`

To fix the error:

- Rename `Style/OpMethods` to `Naming/BinaryOperatorParameter`

This does not look like it is backwards compatible since renaming
`Style/OpMethods` to `Naming/BinaryOperatorParameter` means old versions
do not know about the config for that cop, but it does prevent rubocop
from bumping against an error.

Also, note that the error has been fixed on rubocop's master branch. As
of today that fix has not been pushed to rubygems:
https://github.com/bbatsov/rubocop/commit/e8a61bc118055be131f4d3ae88c6e7e293072818