Ruby: Link to commits and PRs for style items
In 90e1c108 (#473), @jferris added annotations to the Ruby style guide
[README] to help users find the rationale behind decisions, reducing the
effort needed to track down discussions.

Github's online README formatting doesn't create hyperlinks for these
references in the same way that it would for a commit or pull request.
This change wraps those references in links so that it's quicker to get
from the style item to the source contribution.

[README]: https://github.com/thoughtbot/guides/blob/90e1c10/style/ruby/README.md