Clarify meaning of approval
Based on a conversation that happened in Slack on what "approval" means
in the context of code review. This makes it explicit that the PR author
gets to make the final call on whether to merge or make changes. The
reviewer is there to give feedback, not to act as a gatekeeper.

Ideally, discussion and edits go on until both parties are satisfied. In
the event that this drags on too long or no agreement can be reached,
the author may act unilaterally to merge without making some of the
suggested changes.