Switch Hound to use stylelint instead of scss-lint
Based on a [successful Research experiment][research-card], this changes
the Hound configuration in this repo to use stylelint, and disables
scss-lint.

The configuration for stylelint was added in a
[previous commit][stylelint-commit] and matches the scss-lint as much as
possible, so this change should see only minor differences in what
Hound catches.

The reason that it's useful and important to change the Hound config in
this repo, is because we use an organization-wide config in our Hound
account. That means Hound looks to this repo's `.hound.yml` if a project
does not have a local `.hound.yml`. So once this is merged to `master`,
Hound will begin using stylelint in all future thoughtbot PRs, unless
a project declares otherwise.

[research-card]: https://trello.com/c/n17oCxXV/712-replace-scss-lint-with-stylelint
[stylelint-commit]: https://github.com/thoughtbot/guides/commit/ac79b2614a54bae36305a0be77a65c06a7aa910d