Add guideline to avoid JSONB columns
These columns are frequently abused to store has-one relationships when
somebody feels some aversion to adding several columns or a new table.

The actual use case for JSONB columns is slim and specific to storing
raw JSON efficiently. With a known schema, it's better to use separate
columns.

JSONB doesn't support the full range of Postgres types and constraints
(for example, there's no date type in JSON), it's more complicated to
index and query JSONB, it doesn't integrate as well with ORMs, and it's
harder to understand the schema when JSONB is involved.