The thoughtbot Guide to Application Security
This guide is meant to help us all share knowledge and understanding
about what it means to develop an application more securely. It ensures
that we are all on the same page about these important security topics.

The goal here is not to be complete; all of these topics are discussed
in much more depth elsewhere. Rather, the goal is to make sure we have
an overview of what to look for when making software architecture
decisions.