Scaffold a guide for GraphQL (#550)
thoughtbot seems to be doing more work with GraphQL for client projects.
We would like to be to discuss and document what we discover as best
practices, recommended solutions, and issues to avoid for the benefit of
thoughtbot as we go. This commit provides an initial scaffolding for a
guide to GraphQL. Future commits can begin to fill in details.