Document app/services
Many -- perhaps all -- of our projects have an `app/services` directory.
This documents the Platonic ideal of the practice: it is a place for
command objects that orchestrate interactions between two or more
models.

This is in comparison with `app/models`, which holds both ActiveRecord
objects and also data structures related to the app itself.

This is also distinct from `lib`, which is seldom used and often a good
place to move files before extracting them into a gem.

This PR is the result of an internal discussion that happened because of
https://github.com/thoughtbot/suspenders/pull/1015 .