Add CSS Best Practices, move some Sass guides
Commit [9e48ac4] began restructuring our Sass guides. One, perhaps
subtle, change in that commit is a move towards guides on CSS
architecture, not just stylistic ones. And so this commit moves a
portion of those new guides out of the “Style” section of this repo and
in to the “Best Practices” section, where they are better suited.

This commit also moves to describing these as “CSS” guides, not “Sass”
guides. Sass is an extension to CSS; without CSS, Sass doesn’t exist.
Our guides should be written with the opinion that without quality CSS,
quality Sass cannot exist.

[9e48ac4]: https://github.com/thoughtbot/guides/commit/9e48ac4c7f2b35895f069ed559e7f5236afbe755