Offload stylistic CSS/Sass guides to stylelint
We maintain a [sharable stylelint configuration][stylelint-config] that
enforces many of our CSS & Sass stylistic guides. It's hard to maintain
these guides in two places: here in the Guides written as Markdown and
in a stylelint configuration.

Reading guidelines here in this repo as prose is helpful, but consuming
them via stylelint rules is even better, as stylelint can be integrated
into editors, run in Pull Requests via Hound, and has auto-fixing
capabilities.

This commit removes any Markdown-based guides in this repo for things
which are covered by stylelint in our sharable configuration. It also
adds a ‘Best Practice’ to use stylelint.

Below is a list of which stylelint rule covers the guide being removed:

> Use one space between property and value: `width: 20px` not
> `width:20px`.

This is covered by stylelint's `declaration-colon-space-after` rule:
https://stylelint.io/user-guide/rules/declaration-colon-space-after

> Use a blank line above a selector that has styles.

This is covered by stylelint's `rule-empty-line-before` rule:
https://stylelint.io/user-guide/rules/rule-empty-line-before

> Use one space between selector and `{`.

This is covered by stylelint's `block-opening-brace-space-before` rule:
https://stylelint.io/user-guide/rules/block-opening-brace-space-before

> Use double quotation marks.

This is covered by stylelint's `string-quotes` rule:
https://stylelint.io/user-guide/rules/string-quotes

> Use only lowercase, except for hex or string values.

This is mostly covered by these stylelint rules:

- `color-hex-case`:
  https://stylelint.io/user-guide/rules/color-hex-case
- `property-case`:
  https://stylelint.io/user-guide/rules/property-case
- `selector-pseudo-class-case`:
  https://stylelint.io/user-guide/rules/selector-pseudo-class-case
- `selector-pseudo-element-case`:
  https://stylelint.io/user-guide/rules/selector-pseudo-element-case
- `selector-type-case`:
  https://stylelint.io/user-guide/rules/selector-type-case
- `unit-case`:
  https://stylelint.io/user-guide/rules/unit-case
- `value-keyword-case`:
  https://stylelint.io/user-guide/rules/value-keyword-case

> Don't add a unit specification after `0` values, unless required by
> a mixin.

This is covered by stylelint's `length-zero-no-unit` rule:
https://stylelint.io/user-guide/rules/length-zero-no-unit

> Use a leading zero in decimal numbers: `0.5` not `.5`

This is covered by stylelint's `number-leading-zero` rule:
https://stylelint.io/user-guide/rules/number-leading-zero

> Use space around operands: `$variable * 1.5`, not `$variable*1.5`

This is covered by stylelint-scss's `operator-no-unspaced` rule:
https://github.com/kristerkari/stylelint-scss/blob/master/src/rules/operator-no-unspaced/README.md

> Use double colons for pseudo-elements

This is covered by stylelint's `selector-pseudo-element-colon-notation`
rule:
https://stylelint.io/user-guide/rules/selector-pseudo-element-colon-notation

[stylelint-config]: https://github.com/thoughtbot/stylelint-config