Remove factory naming style guidline
This guideline was introduced in
https://github.com/thoughtbot/guides/pull/607. While I agree that naming
factories in this way makes sense, I think it is fairly nitpicky. We
already have "Name variables, methods, and classes to reveal intent" in
our [naming style guidelines], which I think implicitly covers something
like this. I don't care what the name of a factory is as
long as it is contextually meaningful.

[nameing style guidlines]: https://github.com/thoughtbot/guides/tree/master/style#naming