Combine style, best practice, and protocol guides
This change addresses #610.

We remove the arbitrary grouping of guides into "style", "best
practices", and "protocols".

Instead, all guides now live at the top-level of the repository. This
makes it easier to find a guide, and reduces the chance of information
being duplicated between different types of guides for the same topic.

This change also updates the main `README.md` with a grouped table of
contents of all the guides.