Restore Rubocop & Hound configuration (#623)
In #615, we reorganised the repo (for the better!) but this broke Hound
and removed the rubocop.yml we added back in in 9ac4a38. Not every
project is ready to switch to standard and this move would've also broken
Hound's configuration for other languages.

This restores the .rubocop.yml and points Hound to the right files.