Sort ActiveRecord queries by SQL execution order
SQL does not execute the parts of a query in the order it is written.
Instead it follows the [following order]:

* `FROM`/`JOIN` and all the `ON` conditions
* `WHERE`
* `GROUP BY`
* `HAVING`
* `SELECT` (including window functions)
* `ORDER BY`
* `LIMIT`

The idea here is that we can better understand what's happening in a
query if we sort the chain of AR methods in the same order that they
will execute.

Additionally, this order is probably also from "most likely to want to
re-use" down to "most likely to want to break off". It's likely we might
want to have a base scope that does some joining and filtering, but that
other methods might want to control how those results are ordered or how
many get returned.

[following order]: https://jvns.ca/blog/2019/10/03/sql-queries-don-t-start-with-select