Update testing guides to reference system specs
Before, the guides referred to `feature` specs when outlining best
practices for testing with RSpec.

RSpec now uses [`system specs`][rspec-docs] by default, so this updates the
guides to refer to them.

Guidance around the Capybara `feature/scenario` DSL has been removed,
because it is only an [alias for `describe/it`][capybara-docs]. Removing
that specific guidance, reduces the cognitive overhead of remembering
which style to use when writing different types of RSpec test.

[rspec-docs]: https://relishapp.com/rspec/rspec-rails/docs/system-specs/system-spec
[capybara-docs]: https://github.com/teamcapybara/capybara/blob/master/README.md?plain=1#L222-L224