Add paired-programming guideline to share end results (#674)
It's nice to see the results of the code you helped write, because it
helps give a sense of accomplishment, you can review it for more
learning or practice, or you can look for errors. But sometimes you're
pairing on a private repository that only one of you has access to, so
those results aren't automatically available. By explicitly sharing
them, we can have something extra to brag about and learn from.