Update code review guidelines
On my current client project we have been using labels as a prefix to comments we leave on code reviews to clearly display intention and tone. This removes a lot of ambiguity from written text which can be interpreted in many ways.

Our usage of labels has been informed by this article https://conventionalcomments.org/.

This PR would bring this optional way of working to other people within thoughtbot.