Add 2 new examples to remote-meeting guidelines.
This PR adds 2 new examples to the remote-meeting guidelines.

The first example "Meeting Recordings" addresses a common scenario where people want to record a meeting so they themselves can refer back to it or someone who is not present can view it later.

Some video-call software will actively tell you when a meeting is being recorded and seek your permission, others will not. Moreover, someone may be focusing on a different screen.

Therefore, it would be helpful for the person recording the meeting to verbally achieve consensus before recording. 

This will ensure everyone is aware and people can choose to opt out of the meeting or turn their camera off.

The second example "Audio Transcription" is becoming more popular with the advances in AI. There are many 3rd party apps or software which provide this functionality and it can be very useful in many circumstances. 

As an example, for someone with auditory processing challenges, this can allow them to read and even summarize the meeting at a later point in time. Transcriptions can also negate the need for a meeting note-taker.

However, we also need to be sensitive to audio being captured by 3rd party apps. People may not want their words being recorded and stored elsewhere.

Given this, I think it makes sense to let others know if you are using a transcription app so people are aware that what they are saying is being recorded.