Don't use foo, bar, and baz
It is possible that the term foobar has a [questionable
etymology](https://en.wikipedia.org/wiki/Foobar#History_and_etymology),
but in addition, these don't mean anything an in our code, as well as
our documentation, there is almost always a more meaningful example to
use that will better convey intent to other developers, and yourself in
the future.

We already have a line in our guides to "Name variables, methods, and
classes to reveal intent."

However, throughout our open source projects we have used this
placeholder variables in code and documentation when a more meaningful
alternative would have been better.

This adds the point that this guideline also applies to documentation,
and gives the examples of `foo`, `bar`, and `baz`, to hopefully make it
clear what the guideline is saying to not do.