Treat version history as a resource
When composed with care, the version history in a code repository
becomes a valuable resource for the future.

For example, it allows you to use `git blame` to see how files changed
and understand from the commit messages _why_ such changes were made,
allowing you to understand the constraints under what those changes can
be made and explain why the code was written that way.

To do this, you should ensure you write enough detail for someone else
(or perhaps yourself in future) to understand the change in depth. You
should link to external resources, but understand that those can change
and so a link alone is not enough. You should rebase to avoid messy
ordering, and compose your changes with care so that they can be
understood at a glance.