Remove Ember guides
While [Ember] is very cool, these guides haven't been updated in a while
and link to a version of Ember released in 2015.

Some of the recommendations still seem valid, but these guides we
written such a long time ago that I think it would be better to just
start from scratch if we do want to maintain guidance around Ember.

It looks like these guides were [discussed back in 2020] and the concensus
was to remove them, but the PR to do that didn't cover quite all of the
guides.

[Ember]: https://emberjs.com/
[discussed back in 2020]: https://github.com/thoughtbot/guides/issues/584