Use an opinionated set of rules for Rubocop (#722)
* Use an opinionated set of rules for Rubocop

Most projects that thoughtbot is working on currently did not start here.
The "Use standard" recommendation comes from a time that that wasn't
true.

There are a few reasons not to use Standard in existing projects:
- A large project does not already use Standard and retrofitting would be costly
- Need or want Rubocop rules other than Standard's opinionated set

* additions to nme-standard branch (#754)

* Use an opinionated set of rules for Rubocop

Most projects that thoughtbot is working on currently did not start here.
The "Use standard" recommendation comes from a time that that wasn't
true.

There are a few reasons not to use Standard in existing projects:
- A large project does not already use Standard and retrofitting would be costly
- Need or want Rubocop rules other than Standard's opinionated set

* link rubocop guideline to additional information

* fix typos and add trade-offs for using standard

---------

Co-authored-by: Neil Carvalho <me@neil.pro>

---------

Co-authored-by: Valerie Burzynski <valerie.burzynski@thoughtbot.com>