Remove SECURITY.md
Our open source projects include a SECURITY.md file with instructions
about supported versions and how to report a vulnerability. While this
is an open source project, it's not a runnable app, but a collection of
documents.

Recently, as a Markdown linting workflow was added to this project, it
started competing with another workflow that keeps SECURITY.md
up-to-date on how this file should be formatted.

This commit removes both SECURITY.md and the `update-security` GitHub
Actions workflow.